// src/app/core/models/moder_data.ts
export class ModertModel {
    id: string;
    name: string;
    description: string;
    formation: string;
    intensity: string;
    audiance: string;
    min_participants: string;
    max_participants: string;
    video: string;
    avg_rating: number;
    is_validated: boolean;
    validated_by_name: string;
    validated_user_id: string;
    validated_date: string;
    submited_by_name: string;
    submited_by_user_id: string;
    submited_date: Date;
}